+++ lib/active_shipping/carriers/canada_post_pws.rb
@@ -1,5 +1,7 @@
 module ActiveShipping
   class CanadaPostPWS < Carrier
+    
+    cattr_reader :name
     @@name = "Canada Post PWS"
 
     SHIPPING_SERVICES = {
@@ -30,8 +32,6 @@ module ActiveShipping
 
     ENDPOINT = "https://soa-gw.canadapost.ca/"    # production
 
-    SHIPMENT_MIMETYPE = "application/vnd.cpc.ncshipment+xml"
-    RATE_MIMETYPE = "application/vnd.cpc.ship.rate+xml"
     TRACK_MIMETYPE = "application/vnd.cpc.track+xml"
     REGISTER_MIMETYPE = "application/vnd.cpc.registration+xml"
 
@@ -62,11 +62,21 @@ module ActiveShipping
       [:api_key, :secret]
     end
 
+    def self.default_location
+      {
+        :country     => 'CA',
+        :province    => 'ON',
+        :city        => 'Ottawa',
+        :address1    => '61A York St',
+        :postal_code => 'K1N5T2'
+      }
+    end
+
     def find_rates(origin, destination, line_items = [], options = {}, package = nil, services = [])
       url = endpoint + "rs/ship/price"
       request  = build_rates_request(origin, destination, line_items, options, package, services)
-      response = ssl_post(url, request, headers(options, RATE_MIMETYPE, RATE_MIMETYPE))
-      parse_rates_response(response, origin, destination)
+      response = ssl_post(url, request, headers(options, rate_mimetype, rate_mimetype))
+      parse_rates_response(response, origin, destination, !!options[:exclude_tax])
     rescue ActiveUtils::ResponseError, ActiveShipping::ResponseError => e
       error_response(e.response.body, CPPWSRateResponse)
     end
@@ -87,7 +97,7 @@ module ActiveShipping
     # line_items should be a list of PackageItem's
     def create_shipment(origin, destination, package, line_items = [], options = {})
       request_body = build_shipment_request(origin, destination, package, line_items, options)
-      response = ssl_post(create_shipment_url(options), request_body, headers(options, SHIPMENT_MIMETYPE, SHIPMENT_MIMETYPE))
+      response = ssl_post(create_shipment_url(options), request_body, headers(options, shipment_mimetype, shipment_mimetype))
       parse_shipment_response(response)
     rescue ActiveUtils::ResponseError, ActiveShipping::ResponseError => e
       error_response(e.response.body, CPPWSShippingResponse)
@@ -96,12 +106,12 @@ module ActiveShipping
     end
 
     def retrieve_shipment(shipping_id, options = {})
-      response = ssl_post(shipment_url(shipping_id, options), nil, headers(options, SHIPMENT_MIMETYPE, SHIPMENT_MIMETYPE))
+      response = ssl_post(shipment_url(shipping_id, options), nil, headers(options, shipment_mimetype, shipment_mimetype))
       parse_shipment_response(response)
     end
 
     def find_shipment_receipt(shipping_id, options = {})
-      response = ssl_get(shipment_receipt_url(shipping_id, options), headers(options, SHIPMENT_MIMETYPE, SHIPMENT_MIMETYPE))
+      response = ssl_get(shipment_receipt_url(shipping_id, options), headers(options, shipment_mimetype, shipment_mimetype))
       parse_shipment_receipt_response(response)
     end
 
@@ -130,14 +140,14 @@ module ActiveShipping
     end
 
     def find_services(country = nil, options = {})
-      response = ssl_get(services_url(country), headers(options, RATE_MIMETYPE))
+      response = ssl_get(services_url(country), headers(options, rate_mimetype))
       parse_services_response(response)
     rescue ActiveUtils::ResponseError, ActiveShipping::ResponseError => e
       error_response(e.response.body, CPPWSRateResponse)
     end
 
     def find_service_options(service_code, country, options = {})
-      response = ssl_get(services_url(country, service_code), headers(options, RATE_MIMETYPE))
+      response = ssl_get(services_url(country, service_code), headers(options, rate_mimetype))
       parse_service_options_response(response)
     rescue ActiveUtils::ResponseError, ActiveShipping::ResponseError => e
       error_response(e.response.body, CPPWSRateResponse)
@@ -145,7 +155,7 @@ module ActiveShipping
 
     def find_option_details(option_code, options = {})
       url = endpoint + "rs/ship/option/#{option_code}"
-      response = ssl_get(url, headers(options, RATE_MIMETYPE))
+      response = ssl_get(url, headers(options, rate_mimetype))
       parse_option_response(response)
     rescue ActiveUtils::ResponseError, ActiveShipping::ResponseError => e
       error_response(e.response.body, CPPWSRateResponse)
@@ -246,7 +256,7 @@ module ActiveShipping
       line_items = Array(line_items)
 
       builder = Nokogiri::XML::Builder.new do |xml|
-        xml.public_send('mailing-scenario', :xmlns => "http://www.canadapost.ca/ws/ship/rate") do
+        xml.public_send('mailing-scenario', :xmlns => rate_url) do
           customer_number_node(xml, options)
           contract_id_node(xml, options)
           quote_type_node(xml, options)
@@ -261,7 +271,7 @@ module ActiveShipping
       builder.to_xml
     end
 
-    def parse_rates_response(response, origin, destination)
+    def parse_rates_response(response, origin, destination, exclude_tax)
       doc = Nokogiri.XML(response)
       doc.remove_namespaces!
       raise ActiveShipping::ResponseError, "No Quotes" unless doc.at('price-quotes')
@@ -269,7 +279,7 @@ module ActiveShipping
       rates = doc.root.xpath('price-quote').map do |node|
         service_name  = node.at("service-name").text
         service_code  = node.at("service-code").text
-        total_price   = node.at('price-details/due').text
+        total_price   = price_from_node(node, exclude_tax)
         expected_date = expected_date_from_node(node)
         options = {
           :service_code   => service_code,
@@ -280,6 +290,14 @@ module ActiveShipping
         RateEstimate.new(origin, destination, @@name, service_name, options)
       end
       CPPWSRateResponse.new(true, "", {}, :rates => rates)
+    end
+
+    def price_from_node(node, exclude_tax)
+      price = node.at('price-details/due').text
+      return price unless exclude_tax
+      children = node.at('price-details/taxes').children
+      tax_total_cents = children.sum { |node| node.elem? ? Package.cents_from(node.text) : 0 }
+      Package.cents_from(price) - tax_total_cents
     end
 
     # tracking
@@ -339,11 +357,12 @@ module ActiveShipping
     # :cod, :cod_amount, :insurance, :insurance_amount, :signature_required, :pa18, :pa19, :hfp, :dns, :lad
     #
     def build_shipment_request(origin, destination, package, line_items = [], options = {})
-      origin = sanitize_location(origin)
-      destination = sanitize_location(destination)
-
       builder = Nokogiri::XML::Builder.new do |xml|
-        xml.public_send('non-contract-shipment', :xmlns => "http://www.canadapost.ca/ws/ncshipment") do
+        xml.public_send(root_node_name, :xmlns => xmlns_url) do
+          if contract?
+            xml.public_send('transmit-shipment', 'true')
+            xml.public_send('requested-shipping-point', origin.postal_code)
+          end
           xml.public_send('delivery-spec') do
             shipment_service_code_node(xml, options)
             shipment_sender_node(xml, origin, options)
@@ -352,6 +371,7 @@ module ActiveShipping
             shipment_parcel_node(xml, package)
             shipment_notification_node(xml, options)
             shipment_preferences_node(xml, options)
+            shipment_settlement_info_node(xml, options) if contract?
             references_node(xml, options)             # optional > user defined custom notes
             shipment_customs_node(xml, destination, line_items, options)
             # COD Remittance defaults to sender
@@ -365,7 +385,8 @@ module ActiveShipping
       xml.public_send('service-code', options[:service])
     end
 
-    def shipment_sender_node(xml, location, options)
+    def shipment_sender_node(xml, sender, options)
+      location = location_from_hash(sender)
       xml.public_send('sender') do
         xml.public_send('name', location.name)
         xml.public_send('company', location.company) if location.company.present?
@@ -375,13 +396,14 @@ module ActiveShipping
           xml.public_send('address-line-2', location.address2_and_3) unless location.address2_and_3.blank?
           xml.public_send('city', location.city)
           xml.public_send('prov-state', location.province)
-          # xml.public_send('country-code', location.country_code)
+          xml.public_send('country-code', location.country_code) if contract?
           xml.public_send('postal-zip-code', location.postal_code)
         end
       end
     end
 
-    def shipment_destination_node(xml, location, options)
+    def shipment_destination_node(xml, destination, options)
+      location = location_from_hash(destination)
       xml.public_send('destination') do
         xml.public_send('name', location.name)
         xml.public_send('company', location.company) if location.company.present?
@@ -392,7 +414,7 @@ module ActiveShipping
           xml.public_send('city', location.city)
           xml.public_send('prov-state', location.province) unless location.province.blank?
           xml.public_send('country-code', location.country_code)
-          xml.public_send('postal-zip-code', location.postal_code)
+          xml.public_send('postal-zip-code', get_sanitized_postal_code(location))
         end
       end
     end
@@ -419,6 +441,13 @@ module ActiveShipping
       end
     end
 
+    def shipment_settlement_info_node(xml, options)
+      xml.public_send('settlement-info') do
+        xml.public_send('contract-id', '42708517')
+        xml.public_send('intended-method-of-payment', 'Account')
+      end
+    end
+
     def references_node(xml, options)
       # custom values
       # xml.public_send('references') do
@@ -426,7 +455,7 @@ module ActiveShipping
     end
 
     def shipment_customs_node(xml, destination, line_items, options)
-      return unless destination.country_code != 'CA'
+      return unless location_from_hash(destination).country_code != 'CA'
 
       xml.public_send('customs') do
         currency = options[:currency] || "CAD"
@@ -465,27 +494,35 @@ module ActiveShipping
             xml.public_send('width', '%.1f' % ((pkg_dim[1] * 10).round / 10.0)) if pkg_dim.size >= 2
             xml.public_send('height', '%.1f' % ((pkg_dim[0] * 10).round / 10.0)) if pkg_dim.size >= 1
           end
-          xml.public_send('document', false)
+          xml.public_send('document', false) unless contract?
         else
-          xml.public_send('document', true)
+          xml.public_send('document', true) unless contract?
         end
 
-        xml.public_send('mailing-tube', package.tube?)
-        xml.public_send('unpackaged', package.unpackaged?)
+        if !contract?
+          xml.public_send('mailing-tube', package.tube?)
+          xml.public_send('unpackaged', package.unpackaged?)
+        end
       end
     end
 
     def parse_shipment_response(response)
       doc = Nokogiri.XML(response)
       doc.remove_namespaces!
-      raise ActiveShipping::ResponseError, "No Shipping" unless doc.at('non-contract-shipment-info')
+      raise ActiveShipping::ResponseError, "No Shipping" unless doc.at("#{ contract? ? "" : "non-contract-" }shipment-info")
       options = {
         :shipping_id      => doc.root.at('shipment-id').text,
-        :tracking_number  => doc.root.at('tracking-pin').text,
         :details_url      => doc.root.at_xpath("links/link[@rel='details']")['href'],
-        :label_url        => doc.root.at_xpath("links/link[@rel='label']")['href'],
-        :receipt_url      => doc.root.at_xpath("links/link[@rel='receipt']")['href'],
+        :label_url        => doc.root.at_xpath("links/link[@rel='label']")['href']
       }
+      options[:tracking_number] = doc.root.at('tracking-pin').text if doc.root.at('tracking-pin')
+
+      if contract?
+        options[:cost] = doc.root.at_xpath("links/link[@rel='price']")['href']
+      else
+        options[:receipt_url] = doc.root.at_xpath("links/link[@rel='receipt']")['href']
+      end
+
       CPPWSShippingResponse.new(true, "", {}, options)
     end
 
@@ -564,6 +601,36 @@ module ActiveShipping
 
     private
 
+    def contract?
+      @options[:contract_id].present?
+    end
+
+    SHIPMENT_NON_CONTRACT_MIMETYPE = "application/vnd.cpc.ncshipment-v4+xml"
+    SHIPMENT_CONTRACT_MIMETYPE = "application/vnd.cpc.shipment-v7+xml"
+    def shipment_mimetype
+      contract? ? SHIPMENT_CONTRACT_MIMETYPE : SHIPMENT_NON_CONTRACT_MIMETYPE
+    end
+
+    RATE_NON_CONTRACT_MIMETYPE = "application/vnd.cpc.ship.rate+xml"
+    RATE_CONTRACT_MIMETYPE = "application/vnd.cpc.ship.rate-v3+xml"
+    def rate_mimetype
+      contract? ? RATE_CONTRACT_MIMETYPE : RATE_NON_CONTRACT_MIMETYPE
+    end
+
+    RATE_CONTRACT_URL = "http://www.canadapost.ca/ws/ship/rate-v3"
+    RATE_NON_CONTRACT_URL = "http://www.canadapost.ca/ws/ship/rate"
+    def rate_url
+      contract? ? RATE_CONTRACT_URL : RATE_NON_CONTRACT_URL
+    end
+
+    def root_node_name
+      contract? ? 'shipment' : 'non-contract-shipment'
+    end
+
+    def xmlns_url
+      contract? ? "http://www.canadapost.ca/ws/shipment-v7" : "http://www.canadapost.ca/ws/ncshipment-v4"
+    end
+
     def tracking_url(pin)
       case pin.length
         when 12, 13, 16
@@ -577,19 +644,27 @@ module ActiveShipping
 
     def create_shipment_url(options)
       raise MissingCustomerNumberError unless customer_number = options[:customer_number]
-      if @platform_id.present?
-        endpoint + "rs/#{customer_number}-#{@platform_id}/ncshipment"
+      if contract?
+        endpoint + "rs/#{customer_number}/#{customer_number}/shipment"
       else
-        endpoint + "rs/#{customer_number}/ncshipment"
+        if @platform_id.present?
+          endpoint + "rs/#{customer_number}-#{@platform_id}/ncshipment"
+        else
+          endpoint + "rs/#{customer_number}/ncshipment"
+        end
       end
     end
 
     def shipment_url(shipping_id, options = {})
       raise MissingCustomerNumberError unless customer_number = options[:customer_number]
-      if @platform_id.present?
-        endpoint + "rs/#{customer_number}-#{@platform_id}/ncshipment/#{shipping_id}"
+      if contract?
+        endpoint + "rs/#{customer_number}/#{customer_number}/shipment/#{shipping_id}"
       else
-        endpoint + "rs/#{customer_number}/ncshipment/#{shipping_id}"
+        if @platform_id.present?
+          endpoint + "rs/#{customer_number}-#{@platform_id}/ncshipment/#{shipping_id}"
+        else
+          endpoint + "rs/#{customer_number}/ncshipment/#{shipping_id}"
+        end
       end
     end
 
@@ -628,7 +703,7 @@ module ActiveShipping
       }
       headers['Accept'] = accept if accept
       headers['Content-Type'] = content_type if content_type
-      headers['Platform-ID'] = platform_id if platform_id && customer_credentials_valid?(customer_credentials)
+      headers['Platform-ID'] = platform_id if !contract? && platform_id && customer_credentials_valid?(customer_credentials)
       headers
     end
 
@@ -637,7 +712,7 @@ module ActiveShipping
     end
 
     def contract_id_node(xml, options)
-      xml.public_send("contract-id", options[:contract_id]) if options[:contract_id]
+      xml.public_send("contract-id", @options[:contract_id]) if contract?
     end
 
     def quote_type_node(xml, options)
@@ -669,24 +744,25 @@ module ActiveShipping
     end
 
     def origin_node(xml, location)
-      origin = sanitize_location(location)
-      xml.public_send("origin-postal-code", origin.zip)
+      origin = location_from_hash(location)
+      xml.public_send("origin-postal-code", get_sanitized_postal_code(origin))
     end
 
     def destination_node(xml, location)
-      destination = sanitize_location(location)
+      destination = location_from_hash(location)
+      postal_code = get_sanitized_postal_code(destination)
       case destination.country_code
         when 'CA'
           xml.public_send('destination') do
             xml.public_send('domestic') do
-              xml.public_send('postal-code', destination.postal_code)
+              xml.public_send('postal-code', postal_code)
             end
           end
 
         when 'US'
           xml.public_send('destination') do
             xml.public_send('united-states') do
-              xml.public_send('zip-code', destination.postal_code)
+              xml.public_send('zip-code', postal_code)
             end
           end
 
@@ -755,17 +831,14 @@ module ActiveShipping
       DateTime.strptime((options[:shipping_date] || Time.now).to_s, "%Y-%m-%d")
     end
 
-    def sanitize_location(location)
-      location_hash = location.is_a?(Location) ? location.to_hash : location
-      location_hash = sanitize_zip(location_hash)
-      Location.new(location_hash)
+    def location_from_hash(location)
+      return location if location.is_a?(Location)
+      return Location.new(location)
     end
 
-    def sanitize_zip(hash)
-      [:postal_code, :zip].each do |attr|
-        hash[attr].gsub!(/\s+/, '') if hash[attr]
-      end
-      hash
+    def get_sanitized_postal_code(location)
+      return nil if location.nil? || location.postal_code.nil?
+      location.postal_code.gsub(/\s+/, '').upcase
     end
 
     def sanitize_weight_kg(kg)
@@ -813,7 +886,7 @@ module ActiveShipping
   end
 
   class CPPWSTrackingResponse < TrackingResponse
-    DELIVERED_EVENT_CODES = %w(1496 1498 1499 1409 1410 1411 1412 1413 1414 1415 1416 1417 1418 1419 1420 1421 1422 1423 1424 1425 1426 1427 1428 1429 1430 1431 1432 1433 1434 1435 1436 1437 1438)
+    DELIVERED_EVENT_CODES = %w(1408 1409 1410 1411 1412 1413 1414 1415 1416 1417 1418 1419 1420 1421 1422 1423 1424 1425 1426 1427 1428 1429 1430 1431 1432 1433 1434 1435 1436 1437 1438 1441 1442 1496 1497 1498 1499 5300)
     include CPPWSErrorResponse
 
     attr_reader :service_name, :expected_date, :changed_date, :change_reason, :customer_number
@@ -845,13 +918,14 @@ module ActiveShipping
 
   class CPPWSShippingResponse < ShippingResponse
     include CPPWSErrorResponse
-    attr_reader :label_url, :details_url, :receipt_url
+    attr_reader :label_url, :details_url, :receipt_url, :cost
     def initialize(success, message, params = {}, options = {})
       handle_error(message, options)
       super
       @label_url      = options[:label_url]
       @details_url    = options[:details_url]
       @receipt_url    = options[:receipt_url]
+      @cost    = options[:cost]
     end
   end