+++ lib/active_shipping/carriers/usps.rb
@@ -113,11 +113,11 @@ module ActiveShipping
       :package_service => 'PACKAGESERVICE'
     }
 
+    ATTEMPTED_DELIVERY_CODES = %w(02 53 54 55 56 H0)
+
     # Array of U.S. possessions according to USPS: https://www.usps.com/ship/official-abbreviations.htm
     US_POSSESSIONS = %w(AS FM GU MH MP PW PR VI)
 
-    # TODO: figure out how USPS likes to say "Ivory Coast"
-    #
     # Country names:
     # http://pe.usps.gov/text/Imm/immctry.htm
     COUNTRY_NAME_CONVERSIONS = {
@@ -247,8 +247,8 @@ module ActiveShipping
         timestamp = "#{node.at('EventDate').text}, #{node.at('EventTime').text}"
         Time.parse(timestamp)
       else
-        # Arbitrary time in past, because we need to sort properly by time
-        Time.parse("Jan 01, 2000")
+        # Epoch time, because we need to sort properly by time
+        Time.at(0)
       end
 
       event_code = node.at('EventCode').text
@@ -630,8 +630,7 @@ module ActiveShipping
 
         shipment_events = shipment_events.sort_by(&:time)
 
-        # USPS defines a delivery attempt with code 55
-        attempted_delivery_date = shipment_events.detect{ |shipment_event| shipment_event.type_code=="55" }.try(:time)
+        attempted_delivery_date = shipment_events.detect{ |shipment_event| ATTEMPTED_DELIVERY_CODES.include?(shipment_event.type_code) }.try(:time)
 
         if last_shipment = shipment_events.last
           status = last_shipment.status