+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # ActiveShipping CHANGELOG
 
+### v1.8.5
+- Fix UPS TrackResponse parsing for missing elements
+
+### v1.8.4
+- Add price details to rate estimates
+- Fix encoding for UPS responses
+
+### v1.8.3
+- Add description field to rate estimates
+
+### v1.8.2
+- Add option for FedEx label format
+- Fix kunaki remote tests broken due to more shipping options
+
 ### v1.6.1
 - Fix FedEx ShipmentEvents to include event type
 - Skip broken Canada Post remote tests